Move Slack tests into separate files
